Split gadget into independent files
